research_and_dev/iqrah-annotator/backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36

# Data validation (compatible with Python 3.13)
pydantic==2.10.0

# Async support
aiofiles==24.1.0

# Environment variables
python-dotenv==1.0.0
